+++ components/supertype/HISTORY.md
@@ -1,5 +1,6 @@
 ## 1.5.1
 * Fixed problem with list of app names for toClient rules
+* Added objectTemplate.__usageMap__ for list of instantiated templates
 ## 1.5.0
 * Fixed problem in clone where subtypes not being cloned properly
 * Carry over toClient and toServer in intropected properties